trendii-etl/requirements.txt
dbt-postgres==1.9.0
psycopg2-binary==2.9.9
pyarrow==16.1.0
